+++ app/build.gradle.kts
@@ -71,14 +71,9 @@ dependencies {
     androidTestImplementation(libs.androidx.ui.test.junit4)
     debugImplementation(libs.androidx.ui.tooling)
     debugImplementation(libs.androidx.ui.test.manifest)
-
-//    implementation("com.squareup.retrofit2:retrofit:2.9.0")
-//    implementation("com.squareup.retrofit2:converter-gson:2.9.0")
-//    implementation("com.google.accompanist:accompanist-pager:0.28.0")
     implementation("com.google.accompanist:accompanist-pager-indicators:0.28.0")
     implementation("androidx.navigation:navigation-compose:2.7.7")
     implementation("androidx.lifecycle:lifecycle-viewmodel-compose:2.7.0")
-//    implementation("androidx.compose.material3:material3:1.2.1")
     implementation("io.coil-kt:coil-compose:2.6.0")
     implementation("com.google.android.gms:play-services-location:21.3.0") {
         exclude(group = "com.google.guava", module = "listenablefuture")
@@ -91,10 +86,6 @@ dependencies {
     implementation("com.squareup.retrofit2:retrofit:2.11.0")
     implementation("com.squareup.retrofit2:converter-gson:2.11.0")
     implementation("com.squareup.okhttp3:logging-interceptor:4.12.0")
-//    implementation("androidx.compose.material:material-icons-extended:1.6.7")
-//    implementation("io.insert-koin:koin-android:3.5.0")
-//    implementation("io.insert-koin:koin-androidx-compose:3.5.0")
-//    implementation("com.squareup.okhttp3:logging-interceptor:5.0.0-alpha.16")
     implementation("androidx.compose.runtime:runtime-livedata:1.6.8")
     implementation("org.jetbrains.kotlinx:kotlinx-coroutines-play-services:1.7.3")
     implementation("androidx.compose.material:material-icons-extended:1.7.7")
